modify and add project

+++ src/modules/accounts/accounts.service.ts
@@ -1,9 +1,10 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
-import { CreateAccountDto } from './dto/create-account.dto';
+import { ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { UpdateAccountDto } from './dto/update-account.dto';
 import { AccountsRepository } from './accounts.repository';
 import { UsersRepository } from 'src/modules/users/users.repository';
 import { ERROR_MESSAGES } from 'src/common/constants/error-messages';
+import { LogedInUserDto } from '../users/dto/loged-in-user.dto';
+import { CONSTANT } from 'src/common/constants/constant-variable';
 
 @Injectable()
 export class AccountsService {
@@ -13,47 +14,75 @@ export class AccountsService {
   ) {}
 
 
-
-  async create(createAccountDto: CreateAccountDto) {
-    const user = await this.userRepository.findById(createAccountDto.userId) 
+  /* ---------------- Create new accout ---------------- */
+  /* START */
+  async create(userId: string) {
+    const user = await this.userRepository.findById(userId) 
     
     if(!user){
       throw new NotFoundException(ERROR_MESSAGES.USER.NOT_FOUND)
     }
 
     const random10 = Math.floor(1000000000 + Math.random() * 9000000000);
-    createAccountDto.accountNumber = random10.toString();
+    const accountNumber = random10.toString();
 
-    return this.accountsRepository.createAccount(createAccountDto);
+    return this.accountsRepository.createAccount(userId, accountNumber);
   }
+  /* END */
 
 
-
+  /* ---------------- Find all existing account ---------------- */
   findAll() {
-    return `This action returns all accounts`;
+    return this.accountsRepository.findAll()
   }
 
 
-
+  /* ---------------- Find all loged in user account ---------------- */
   findAllByUserId(userId: string) {
     return this.accountsRepository.getAccountByUserId(userId)
   }
 
 
+  /* ---------------- Find one account ---------------- */
+  /* START */
+  async findOne(user: LogedInUserDto, accountNumber: string) {
+    const account = await this.accountsRepository.getAccountByAccountNumber(accountNumber)
 
-  findOne(accountNumber: string) {
-    return this.accountsRepository.getAccountByAccountNumber(accountNumber)
+    if(!account){
+      throw new NotFoundException(ERROR_MESSAGES.ACCOUNT.NOT_FOUND + accountNumber)
+    }
+
+    if(!(user.id === account?.userId)){
+      throw new ForbiddenException(ERROR_MESSAGES.USER.VIEW_FORBIDDEN)
+    }
+
+    return account
   }
+  /* END */
 
 
+  /* ---------------- Update account ---------------- */
+  async update(updateAccountDto: UpdateAccountDto) {
+    const account = await this.accountsRepository.getAccountByAccountNumber(updateAccountDto.accountNumber)
+    if(!account){
+      throw new NotFoundException(ERROR_MESSAGES.ACCOUNT.NOT_FOUND + updateAccountDto.accountNumber)
+    }
 
-  update(updateAccountDto: UpdateAccountDto) {
     return this.accountsRepository.updateAccount(updateAccountDto);
   }
-
-
   
-  remove(accountNumber: string) {
+
+  /* ---------------- Delete account ---------------- */
+  async remove(user: LogedInUserDto, accountNumber: string) {
+    const account = await this.accountsRepository.getAccountByAccountNumber(accountNumber)
+    if(!account){
+      throw new NotFoundException(ERROR_MESSAGES.ACCOUNT.NOT_FOUND + accountNumber)
+    }
+
+    if(!(account.userId === user.id) || user.role === CONSTANT.ROLE.ADMIN){
+      throw new ForbiddenException(ERROR_MESSAGES.ACCOUNT.DELETE_FORBIDDEN)
+    }
+
     return this.accountsRepository.deleteAccount(accountNumber)
   }
 }
